backend/requirements.txt
fastapi>=0.104.0
python-socketio>=5.9.0
uvicorn>=0.24.0
httpx>=0.25.0
pydantic>=2.4.2
python-multipart>=0.0.6
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
jinja2>=3.1.2